restructuring, generate_edge_bundle -> .protocol.edge, sha256_hash -> rid_lib.ext.utils, .protocol.secure_models -> .protocol.envelope
