+++ src/koi_net/protocol/edge.py
@@ -1,7 +1,8 @@
 from enum import StrEnum
 from pydantic import BaseModel
 from rid_lib import RIDType
-from rid_lib.types import KoiNetNode
+from rid_lib.ext.bundle import Bundle
+from rid_lib.types import KoiNetEdge, KoiNetNode
 
 
 class EdgeStatus(StrEnum):
@@ -18,3 +19,24 @@ class EdgeProfile(BaseModel):
     edge_type: EdgeType
     status: EdgeStatus
     rid_types: list[RIDType]
+
+
+def generate_edge_bundle(
+    source: KoiNetNode,
+    target: KoiNetNode,
+    rid_types: list[RIDType],
+    edge_type: EdgeType
+) -> Bundle:
+    edge_rid = KoiNetEdge.generate(source, target)
+    edge_profile = EdgeProfile(
+        source=source,
+        target=target,
+        rid_types=rid_types,
+        edge_type=edge_type,
+        status=EdgeStatus.PROPOSED
+    )
+    edge_bundle = Bundle.generate(
+        edge_rid,
+        edge_profile.model_dump()
+    )
+    return edge_bundle
