+++ src/koi_net/protocol/envelope.py
